+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - chore: Add ruby-lsp gem
+- refactor: Parse JSON with symbolize names
 
 ## [v2.1.0] - 2024-01-19